+++ brres/model.py
@@ -1,6 +1,65 @@
 # ---------------------------------------------------------------------
 #   Model class
 # ---------------------------------------------------------------------
+from material import *
+from layer import *
+from shader import *
+from struct import *
+from pack import *
+from unpack import *
+from fileobject import *
+from brres import *
+
+class IndexGroup:
+    structs = {
+        "h" : Struct("> I I"),
+        "indexGroup" : Struct("> H H H H I I"),
+    }
+    def __init__(self, file, offset):
+        if offset:
+            file.offset = offset
+        self.isModified = False
+        self.offset = file.offset
+        self.h = file.read(self.structs["h"], 8)
+        # print("Group byte len, numEntries: {}".format(self.h))
+        self.entries = []
+        self.entryNames = []
+        self.entryOffsets = []
+        # file.offset += 16
+        for i in range(self.h[1] + 1):
+            entry = file.read(self.structs["indexGroup"], 16)
+            # if self.h[1] == 3:
+            # print("{} Entry id, uk, left, right, namep, datap {}".format(file.offset - 16, entry))
+            if i != 0:
+                name = file.unpack_name(entry[4] + self.offset)
+                self.entryNames.append(name)
+                self.entryOffsets.append(entry[5] + self.offset)
+            # else:
+            #     self.entryNames.append("Null")
+
+            self.entries.append(entry)
+
+    def updateEntryOffset(self, offset, index):
+        self.entryOffsets[index] = offset
+        self.isModified = True
+
+    # does not handle changing left right id etc
+    def repack(self, file):
+        if not self.isModified:
+            return
+        offset = self.offset
+        # pack header
+        pack_into("> I I", file.file, offset, self.h[0], self.h[1])
+        offset += 8
+        # pack entries
+        for i in range(len(self.entries)):
+            entry = self.entries[i]
+            entryOffset = entry[5] if i == 0 else self.entryOffsets[i - 1] - self.offset
+            # print("Offset is {} for entry {}".format(self.offset + entryOffset, self.entries[i]))
+            pack_into("> 4H 2I", file.file, offset, entry[0], entry[1], entry[2],
+            entry[3], entry[4], entryOffset)
+            offset += 16
+
 
 class Model:
     def __init__(self, file, subFileHeader):
